requirements.txt
asgiref==3.8.1
beautifulsoup4==4.13.4
bs4==0.0.2
certifi==2025.8.3
charset-normalizer==3.4.3
Django==5.1.3
django-anymail==13.1
django-browser-reload==1.17.0
djangorestframework==3.15.2
idna==3.10
mysqlclient==2.2.6
pillow==11.0.0
python-dotenv==1.0.1
requests==2.32.4
soupsieve==2.7
sqlparse==0.5.2
typing_extensions==4.14.1
tzdata==2024.2
urllib3==2.5.0
